unify operations
