+++ rollingops/tests/integration/charms/common.py
@@ -37,7 +37,7 @@ logger = logging.getLogger(__name__)
 TRACE_FILE = pathops.LocalPath('/var/lib/charm-rolling-ops/transitions.log')
 
 
-def now_timestamp_str() -> str:
+def _now_timestamp_str() -> str:
     """UTC timestamp as a string using ISO 8601 format."""
     return datetime.now(UTC).isoformat()
 
@@ -122,7 +122,7 @@ class Charm(CharmBase):
     def _record_transition(self, name: str, **data: Any) -> None:
         TRACE_FILE.parent.mkdir(parents=True, exist_ok=True)
         payload = {
-            'ts': now_timestamp_str(),
+            'ts': _now_timestamp_str(),
             'unit': self.model.unit.name,
             'event': name,
             **data,
